+++ tests/FregataTest.php
@@ -50,4 +50,38 @@ class FregataTest extends TestCase
         self::expectException(ConnectionException::class);
         $fregata->addMigrator($migrator);
     }
+
+    /**
+     * There must be at least 1 registered migrator
+     */
+    public function testRunWithoutMigrator()
+    {
+        $fregata = new Fregata();
+
+        self::expectException(\LogicException::class);
+        $fregata->run();
+    }
+
+    /**
+     * Run the migration with a single migrator
+     */
+    public function testRunWithMigrator()
+    {
+        $source = new class extends AbstractConnection {
+            public string $url = 'mysql://root:root@127.0.0.1:3306/fregata_source';
+        };
+        $target = new class extends AbstractConnection {
+            public string $url = 'pgsql://postgres:postgres@127.0.0.1:5432/fregata_target';
+        };
+
+        $migrator = $this->createMock(MigratorInterface::class);
+        $migrator->method('getSourceConnection')->willReturn(get_class($source));
+        $migrator->method('getTargetConnection')->willReturn(get_class($target));
+
+        $fregata = (new Fregata())
+            ->addMigrator($migrator);
+
+        $result = $fregata->run();
+        self::assertNull($result);
+    }
 }